Split off Multiboot information parsing into information module
